add log pipeline multi pipeline e2e test, refactor trace and metric multi pipeline e2e tests
