+++ test/e2e/logs_multi_pipeline_test.go
@@ -9,15 +9,14 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	kitk8s "github.com/kyma-project/telemetry-manager/test/testkit/k8s"
-	kitkyma "github.com/kyma-project/telemetry-manager/test/testkit/kyma"
 	"github.com/kyma-project/telemetry-manager/test/testkit/mocks/backend"
-	"github.com/kyma-project/telemetry-manager/test/testkit/mocks/telemetrygen"
+	"github.com/kyma-project/telemetry-manager/test/testkit/mocks/loggen"
 	"github.com/kyma-project/telemetry-manager/test/testkit/suite"
 	"github.com/kyma-project/telemetry-manager/test/testkit/verifiers"
 )
 
-var _ = Describe(suite.ID(), Label(suite.LabelMetrics), Ordered, func() {
-	Context("When multiple metricpipelines exist", Ordered, func() {
+var _ = Describe(suite.ID(), Label(suite.LabelLogs), Ordered, func() {
+	Context("When multiple logpipelines exist", Ordered, func() {
 		var (
 			mockNs            = suite.IDWithSuffix("multi-pipeline")
 			backend1Name      = suite.IDWithSuffix("backend-1")
@@ -32,23 +31,28 @@ var _ = Describe(suite.ID(), Label(suite.LabelMetrics), Ordered, func() {
 			var objs []client.Object
 			objs = append(objs, kitk8s.NewNamespace(mockNs).K8sObject())
 
-			backend1 := backend.New(mockNs, backend.SignalTypeMetrics, backend.WithName(backend1Name))
+			backend1 := backend.New(mockNs, backend.SignalTypeLogs, backend.WithName(backend1Name))
 			objs = append(objs, backend1.K8sObjects()...)
 			backend1ExportURL = backend1.ExportURL(proxyClient)
 
-			metricPipeline1 := kitk8s.NewMetricPipelineV1Alpha1(pipeline1Name).WithOutputEndpointFromSecret(backend1.HostSecretRefV1Alpha1())
-			objs = append(objs, metricPipeline1.K8sObject())
+			logPipeline1 := kitk8s.NewLogPipelineV1Alpha1(pipeline1Name).
+				WithSecretKeyRef(backend1.HostSecretRefV1Alpha1()).
+				WithHTTPOutput().
+				Persistent(suite.IsOperational())
+			objs = append(objs, logPipeline1.K8sObject())
 
-			backend2 := backend.New(mockNs, backend.SignalTypeMetrics, backend.WithName(backend2Name))
+			backend2 := backend.New(mockNs, backend.SignalTypeLogs, backend.WithName(backend2Name))
+			logProducer := loggen.New(mockNs)
 			objs = append(objs, backend2.K8sObjects()...)
+			objs = append(objs, logProducer.K8sObject())
 			backend2ExportURL = backend2.ExportURL(proxyClient)
 
-			metricPipeline := kitk8s.NewMetricPipelineV1Alpha1(pipeline2Name).WithOutputEndpointFromSecret(backend2.HostSecretRefV1Alpha1())
-			objs = append(objs, metricPipeline.K8sObject())
+			logPipeline2 := kitk8s.NewLogPipelineV1Alpha1(pipeline2Name).
+				WithSecretKeyRef(backend2.HostSecretRefV1Alpha1()).
+				WithHTTPOutput().
+				Persistent(suite.IsOperational())
+			objs = append(objs, logPipeline2.K8sObject())
 
-			objs = append(objs,
-				telemetrygen.New(mockNs, telemetrygen.SignalTypeMetrics).K8sObject(),
-			)
 			return objs
 		}
 
@@ -62,22 +66,19 @@ var _ = Describe(suite.ID(), Label(suite.LabelMetrics), Ordered, func() {
 		})
 
 		It("Should have running pipelines", func() {
-			verifiers.MetricPipelineShouldBeHealthy(ctx, k8sClient, pipeline1Name)
-			verifiers.MetricPipelineShouldBeHealthy(ctx, k8sClient, pipeline2Name)
+			verifiers.LogPipelineShouldBeHealthy(ctx, k8sClient, pipeline1Name)
+			verifiers.LogPipelineShouldBeHealthy(ctx, k8sClient, pipeline2Name)
 		})
 
-		It("Should have a running metric gateway deployment", func() {
-			verifiers.DeploymentShouldBeReady(ctx, k8sClient, kitkyma.MetricGatewayName)
-		})
-
-		It("Should have a metrics backend running", func() {
+		It("Should have a log backend running", func() {
 			verifiers.DeploymentShouldBeReady(ctx, k8sClient, types.NamespacedName{Name: backend1Name, Namespace: mockNs})
 			verifiers.DeploymentShouldBeReady(ctx, k8sClient, types.NamespacedName{Name: backend2Name, Namespace: mockNs})
 		})
 
-		It("Should deliver telemetrygen metrics", func() {
-			verifiers.MetricsFromNamespaceShouldBeDelivered(proxyClient, backend1ExportURL, mockNs, telemetrygen.MetricNames)
-			verifiers.MetricsFromNamespaceShouldBeDelivered(proxyClient, backend2ExportURL, mockNs, telemetrygen.MetricNames)
+		It("Should have produced logs in the backend", func() {
+			verifiers.LogsShouldBeDelivered(proxyClient, loggen.DefaultName, backend1ExportURL)
+			verifiers.LogsShouldBeDelivered(proxyClient, loggen.DefaultName, backend2ExportURL)
 		})
 	})
+
 })
